Changed things, see description
Deleted uneeded functions in cogsplanner, created a "start" functions that gets the user to input the courses they've taken. Added a function that counts the number of credits the user has taken. Some functions to add would be one that determines how many credits the user needs to take to complete their degree. A function that just spits out basic information about the COGS comp sci degree and what not.